Integrate ActivityExportingDelegate's code into AnkiActivity superclass
This allows removing the exporting code from top level activities like
DeckPicker or CardBrowser.
