+++ AnkiDroid/src/main/java/com/ichi2/anki/CardBrowser.kt
@@ -91,14 +91,10 @@ import com.ichi2.anki.dialogs.DeckSelectionDialog
 import com.ichi2.anki.dialogs.DeckSelectionDialog.Companion.newInstance
 import com.ichi2.anki.dialogs.DeckSelectionDialog.DeckSelectionListener
 import com.ichi2.anki.dialogs.DeckSelectionDialog.SelectableDeck
-import com.ichi2.anki.dialogs.ExportReadyDialog.Companion.KEY_EXPORT_PATH
-import com.ichi2.anki.dialogs.ExportReadyDialog.Companion.REQUEST_EXPORT_SAVE
-import com.ichi2.anki.dialogs.ExportReadyDialog.Companion.REQUEST_EXPORT_SHARE
 import com.ichi2.anki.dialogs.SimpleMessageDialog
 import com.ichi2.anki.dialogs.tags.TagsDialog
 import com.ichi2.anki.dialogs.tags.TagsDialogFactory
 import com.ichi2.anki.dialogs.tags.TagsDialogListener
-import com.ichi2.anki.export.ActivityExportingDelegate
 import com.ichi2.anki.export.ExportDialogFragment
 import com.ichi2.anki.model.CardStateFilter
 import com.ichi2.anki.model.CardsOrNotes
@@ -182,7 +178,6 @@ open class CardBrowser :
     private var mySearchesItem: MenuItem? = null
     private var previewItem: MenuItem? = null
     private var undoSnackbar: Snackbar? = null
-    private lateinit var exportingDelegate: ActivityExportingDelegate
 
     // card that was clicked (not marked)
     override var currentCardId
@@ -377,7 +372,6 @@ open class CardBrowser :
             return
         }
         tagsDialogFactory = TagsDialogFactory(this).attachToActivity<TagsDialogFactory>(this)
-        exportingDelegate = ActivityExportingDelegate(this)
         super.onCreate(savedInstanceState)
         if (!ensureStoragePermissions()) {
             return
@@ -441,8 +435,6 @@ open class CardBrowser :
 
         startLoadingCollection()
 
-        exportingDelegate.onRestoreInstanceState(savedInstanceState)
-
         // Selected cards aren't restored on activity recreation,
         // so it is necessary to dismiss the change deck dialog
         getCurrentDialogFragment<DeckSelectionDialog>()?.let { dialogFragment ->
@@ -473,18 +465,6 @@ open class CardBrowser :
                     showSnackbar(TR.browsingNotesUpdated(count))
                 }
             }
-        }
-
-        supportFragmentManager.setFragmentResultListener(REQUEST_EXPORT_SAVE, this) { _, bundle ->
-            exportingDelegate.saveExportFile(
-                bundle.getString(KEY_EXPORT_PATH) ?: error("Missing required exportPath!"),
-            )
-        }
-
-        supportFragmentManager.setFragmentResultListener(REQUEST_EXPORT_SHARE, this) { _, bundle ->
-            exportingDelegate.shareFile(
-                bundle.getString(KEY_EXPORT_PATH) ?: error("Missing required exportPath!"),
-            )
         }
     }
 
@@ -1638,7 +1618,6 @@ open class CardBrowser :
     public override fun onSaveInstanceState(outState: Bundle) {
         // Save current search terms
         outState.putString("mSearchTerms", viewModel.searchTerms)
-        exportingDelegate.onSaveInstanceState(outState)
         super.onSaveInstanceState(outState)
     }
